requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.10.1
django-on-heroku==1.1.2
django-rest-framework==0.1.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Pillow==9.0.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
